+++ include/qalgorithms_qpeaks.h
@@ -265,9 +265,9 @@ namespace qAlgorithms
     // ### pre-calculate the regression matrix ### //
 #define MAXSCALE 63
 
-    constexpr std::array<float, (MAXSCALE + 1) * 6> initialize()
+    constexpr std::array<double, (MAXSCALE + 1) * 6> initialize()
     { // array to store the 6 unique values of the inverse matrix for each scale
-        std::array<float, (MAXSCALE + 1) * 6> invArray = {0};
+        std::array<double, (MAXSCALE + 1) * 6> invArray = {0};
         // init invArray
         // XtX = transposed(Matrix X ) * Matrix X
         // XtX_xy: x = row number; y = column number
